Change in Code Structure, epic Refactoring... (#10)
* 💄 EPIC: Code structure changed, Refactoring. PHASE 1

* 📖 Few updates in README.md
