+++ README.md
@@ -18,7 +18,7 @@
 <br /><br />
 
 # 📦 `bulk-mail-cli`
-### A mini-MailChimp in your terminal!
+### The free MailChimp in your terminal!
 
 <br /><br />
 
@@ -85,30 +85,27 @@ Just like what you see below:
 $ bulkmail mail
 
 Please fill in the needed data...
-service: gmail/outlook/aol/webmail/hotmail/pop3
+service: gmail/yahoo/custom
 email: yourEmail@domain.com
 password: *************
-tls: <will be configured according to what service you choose>
 
 Wow! You are connected 😀
 
 Path to CSV: <you can use bash's auto navigating feature>
 Path to Mail Template: <you can use bash's auto navigating feature>
 
+From name: John Doe
+
 There you go!.. Now, it is all my task!
 
 Bulk Mail in progress:
-[=========================>] 100%
+[-------------------------------------------------->] 100%
 
 Mails sent successfully in 9.43 seconds.
 
 Thank you for using bulk-mail-cli 😀
 ```
-It will ask for credentials only for the first time. 
-
-_Credentials are stored locally in your machine. The CLI is no where connected to cloud or servers to track the information entered._
-
-**THIS PROJECT IS MORE THAN SECURE FOR YOU TO WORK WITH.**
+**Credentials are asked only for the first time.**
 
 <br />
 
@@ -137,7 +134,12 @@ If the CSV file has emails fed in some other coulumn name (such as `user-email`,
 
 <br />
 
-**AS PER NOW, THESE FEATURES ARE UNDER DEVELOPMENT. WANNA CONTRIBUTE?**
+## Available in 10 major languages
+```
+$ bulkmail config --lang
+```
+
+**FEW MORE THINGS ARE STILL UNDER DEVELOPMENT. WANNA CONTRIBUTE?**
 
 <br /><br />
