+++ src/assets/Bulkmail_demo/CSV_FILE.csv
@@ -1,0 +1,1 @@
+Hi There!
