+++ src/assets/Bulkmail_demo/TEMPLATE_FILE.html
@@ -1,0 +1,1 @@
+Hello World!
