+++ src/lib/commands/tools/changeUsername.util.js
@@ -1,6 +1,6 @@
 import { terminal } from 'terminal-kit'
-import BulkMailCli_settings from '../../settings/settings.util'
-import BulkMailCli_i18n from '../../i18n/i18n.util'
+import BulkMailCli_settings from '../../settings'
+import BulkMailCli_i18n from '../../i18n'
 
 var { setSetting, getSetting } = BulkMailCli_settings
 var { getText } = BulkMailCli_i18n
@@ -29,4 +29,4 @@ var changeUsername = async () => {
 
 }
 
-export default changeUsername
+export { changeUsername }
