+++ src/lib/commands/tools/index.js
@@ -1,0 +1,6 @@
+export * from './help.util'
+export * from './version.util'
+export * from './demo.util'
+export * from './authSession.util'
+export * from './changeUsername.util'
+export * from './selectLang.util'
