+++ src/lib/utilities/checkConnection.util.js
