+++ src/lib/utilities/index.js
@@ -1,0 +1,4 @@
+export * from './checkConnection.util'
+export * from './copyFile.util'
+export * from './minimist.util'
+export * from './writeFile.util'
